+++ Event/xAOD/xAODTrigL1Calo/xAODTrigL1Calo/versions/CMXEtSumsAuxContainer_v1.h
@@ -4,7 +4,7 @@
   Copyright (C) 2002-2017 CERN for the benefit of the ATLAS collaboration
 */
 
-// $Id: CMXEtSumsAuxContainer_v1.h 663426 2015-04-28 11:45:23Z amazurov $
+// $Id: CMXEtSumsAuxContainer_v1.h 687949 2015-08-06 15:48:49Z amazurov $
 #ifndef XAODTRIGL1CALO_VERSIONS_CMXETSUMSAUXCONTAINER_V1_H
 #define XAODTRIGL1CALO_VERSIONS_CMXETSUMSAUXCONTAINER_V1_H
 
@@ -22,8 +22,8 @@ namespace xAOD{
   ///
   /// @author John Morris <john.morris@cern.ch>
   ///
-  /// $Revision: 663426 $
-  /// $Date: 2015-04-28 13:45:23 +0200 (Tue, 28 Apr 2015) $
+  /// $Revision: 687949 $
+  /// $Date: 2015-08-06 17:48:49 +0200 (Thu, 06 Aug 2015) $
 
   class CMXEtSumsAuxContainer_v1 : public AuxContainerBase {
     public:
@@ -37,9 +37,9 @@ namespace xAOD{
       std::vector<std::vector<uint16_t> > etVec;
       std::vector<std::vector<uint16_t> > exVec;
       std::vector<std::vector<uint16_t> > eyVec;
-      std::vector<std::vector<uint8_t> > etErrorVec;
-      std::vector<std::vector<uint8_t> > exErrorVec;
-      std::vector<std::vector<uint8_t> > eyErrorVec;
+      std::vector<std::vector<uint32_t> > etErrorVec;
+      std::vector<std::vector<uint32_t> > exErrorVec;
+      std::vector<std::vector<uint32_t> > eyErrorVec;
 
   }; // class CMXEtSumsAuxContainer_v1
 } // namespace xAOD
